pg_stat_statements will build
