+++ cli/src/commands/build.rs
@@ -1,3 +1,4 @@
+use std::fs;
 use super::SubCommand;
 use crate::commands::generic_build::build_generic;
 use crate::commands::pgrx::build_pgrx;
@@ -40,6 +41,7 @@ impl SubCommand for BuildCommand {
                 if self.version.is_some() || self.name.is_some() {
                     return Err(anyhow!("--version and --name are collected from Cargo.toml when building pgrx extensions, please do not configure"));
                 }
+
                 build_pgrx(
                     self.dockerfile_path.clone(),
                     self.platform.clone(),
@@ -63,10 +65,30 @@ impl SubCommand for BuildCommand {
                     "--version and --name are required when building a makefile based extension"
                 ));
             }
+            let mut dockerfile = String::new();
+            if self.dockerfile_path.clone().is_some() {
+                let dockerfile_path_unwrapped = self.dockerfile_path.clone().unwrap();
+                println!("Using Dockerfile at {}", &dockerfile_path_unwrapped);
+                dockerfile = fs::read_to_string(dockerfile_path_unwrapped.as_str())?;
+            } else {
+                dockerfile = include_str!("./builders/Dockerfile.generic").to_string();
+            }
+
+            let mut install_command_split: Vec<&str> = vec![];
+            if let Some(install_command) = self.install_command.as_ref() {
+                install_command_split.push("/bin/sh");
+                install_command_split.push("-c");
+                install_command_split.push(install_command);
+            } else {
+                install_command_split = vec!["make", "install"];
+            }
+            println!("Using install command {}", install_command_split.clone().join(" "));
+
+            let dockerfile = dockerfile.as_str();
             build_generic(
-                self.dockerfile_path.clone(),
+                dockerfile,
                 self.platform.clone(),
-                vec!["make", "install"],
+                install_command_split,
                 path,
                 &self.output_path,
                 self.name.clone().unwrap().as_str(),
