+++ cli/tests/.gitignore
@@ -1,2 +1,6 @@
 pg_tle
 pg_tle/**
+postgres/**
+postgres
+.trunk
+.trunk/**
