+++ cli/tests/Dockerfile.pg_stat_statements
@@ -1,0 +1,15 @@
+ARG PG_VERSION=15
+
+FROM quay.io/coredb/c-builder:pg${PG_VERSION}
+
+ARG PG_VERSION=15
+
+USER root
+
+# Postgres build dependencies
+# https://wiki.postgresql.org/wiki/Compile_and_Install_from_source_code
+RUN apt-get update && apt-get install -y  build-essential libreadline-dev zlib1g-dev flex bison libxml2-dev libxslt-dev libssl-dev libxml2-utils xsltproc ccache
+
+COPY --chown=postgres:postgres . .
+RUN ./configure
+RUN cd contrib/pg_stat_statements && make
